+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## (6.1.0) IN PROGRESS
 
+* [UITEN-169](https://issues.folio.org/browse/UITEN-169) Move Tenant/Bursar exports to Users/Transfer criteria
 * [UITEN-65](https://issues.folio.org/browse/UITEN-65) Disable key-bindings; it was just a mean tease.
 * [UITEN-165](https://issues.folio.org/browse/UITEN-165) Add remote storage returning workflow details to location view.
 * [UITEN-174](https://issues.folio.org/browse/UITEN-174) Update "remote-storage-mappings" interface